requirements.txt
fastapi==0.123.4
uvicorn==0.30.0
sqlmodel==0.0.18
sqlalchemy==2.1.5
psycopg2-binary==2.9.11
python-jose==3.5.0
passlib==1.7.4
bcrypt==3.2.2
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
fastapi-mail==1.6.0
email-validator==2.3.0
python-multipart==0.0.20
asyncpg==0.31.0
anyio==4.12.0
cryptography==46.0.3
